+++ docs/source/release-notes.rst
@@ -11,6 +11,15 @@ v1.9 Notes
 
 v1.9.3 (2024-04-XX)
 -------------------
+Development of connectivity manager `ConnMan`:
+
+- Add case `ieee14_conn.xlsx` for demonstration.
+- Add `ConnMan` class to manage connectivity.
+- Add `ConnMan` to `System` to as an attribute `conn`.
+- Add a demo notebook for `ConnMan`.
+
+Other changes:
+
 - In the ``dae`` module, change `self.t.itemset` to array assignment to ensure compatibility with NumPy 2.0.
 - Follow RTD's deprecation of Sphinx context injection at build time
 - In symbolic processor, most variables are assumed to be real, except some
@@ -22,6 +31,7 @@ v1.9.3 (2024-04-XX)
 - Add parameter `allow_all=False` to `ModelData.find_idx()` `GroupBase.find_idx()` to allow searching all matches.
 - Add method `GroupBase.get_all_idxes()` to get all indices of a group.
 - Enhanced three-winding transformer parsing in PSS/E raw files by assigning the equivalent star bus `area`, `owner`, and `zone` using the high-voltage bus values.
+- Specify `multiprocess <=0.70.16` in requirements as 0.70.17 does not support Linux.
 
 v1.9.2 (2024-03-25)
 -------------------